+++ x/foundation/expected_keepers.go
@@ -21,6 +21,8 @@ type (
 		SendCoinsFromModuleToModule(ctx sdk.Context, senderModule string, recipientModule string, amt sdk.Coins) error
 		SendCoinsFromModuleToAccount(ctx sdk.Context, senderModule string, recipientAddr sdk.AccAddress, amt sdk.Coins) error
 		SendCoinsFromAccountToModule(ctx sdk.Context, senderAddr sdk.AccAddress, recipientModule string, amt sdk.Coins) error
+
+		MintCoins(ctx sdk.Context, moduleName string, amt sdk.Coins) error
 	}
 
 	// StakingKeeper defines the staking module interface contract needed by the