+++ x/foundation/genesis.go
@@ -7,6 +7,10 @@ import (
 	sdkerrors "github.com/line/lbm-sdk/types/errors"
 )
 
+const (
+	GovMintMaxCount = 1
+)
+
 // DefaultGenesisState creates a default GenesisState object
 func DefaultGenesisState() *GenesisState {
 	return &GenesisState{
@@ -109,6 +113,10 @@ func ValidateGenesis(data GenesisState) error {
 
 	if err := data.Pool.ValidateBasic(); err != nil {
 		return err
+	}
+
+	if data.GovMintLeftCount > GovMintMaxCount {
+		return ErrInvalidGovMintLeftCount.Wrapf("invalid value: %d, max: %d", data.GovMintLeftCount, GovMintMaxCount)
 	}
 
 	return nil