+++ x/foundation/msgs.go
@@ -353,3 +353,28 @@ func (m MsgRevoke) GetSigners() []sdk.AccAddress {
 	signer := sdk.MustAccAddressFromBech32(m.Operator)
 	return []sdk.AccAddress{signer}
 }
+
+var _ sdk.Msg = (*MsgGovMint)(nil)
+
+// ValidateBasic implements Msg.
+func (m MsgGovMint) ValidateBasic() error {
+	if _, err := sdk.AccAddressFromBech32(m.Operator); err != nil {
+		return sdkerrors.ErrInvalidAddress.Wrapf("invalid operator address: %s", m.Operator)
+	}
+
+	if m.Amount.Empty() {
+		return sdkerrors.ErrInvalidCoins.Wrapf("The gov-mint request amount is empty.")
+	}
+
+	if !m.Amount.IsValid() {
+		return sdkerrors.ErrInvalidCoins.Wrap(m.Amount.String())
+	}
+
+	return nil
+}
+
+// GetSigners implements Msg.
+func (m MsgGovMint) GetSigners() []sdk.AccAddress {
+	signer := sdk.MustAccAddressFromBech32(m.Operator)
+	return []sdk.AccAddress{signer}
+}